requirements.txt
pydantic>=2.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
streamlit>=1.35.0
httpx>=0.27.0
openai>=1.30.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
loguru>=0.7.0
tavily-python>=0.5.0
chromadb>=0.5.0
